requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
mecab-python3==1.0.10
requests==2.32.3
unidic-lite==1.0.8
urllib3==2.2.3
